Graphics Library Interface 제작중
Graphics Engine Solution 솔루션 추가.
D11 라이브러리 프로젝트 폴더 생성.

+++ DUOLGraphicsLibrary/include/DUOLGraphicsLibrary/EntityBase.h
@@ -13,7 +13,7 @@ namespace DUOLGraphicsLibrary
 	public:											\
 		inline static int GetEntityID()				\
 		{											\
-			return ID;								\
+			return static_cast<int>(ID);			\
 		}											\
 		bool IsInstanceOf(int id) const override
